Package and split job details components
