+++ internal/lookout/ui/src/components/job-details/SchedulingHistory.css
@@ -1,0 +1,15 @@
+.scheduling-history-title {
+    margin-top: 2em;
+    margin-left: 16px;
+    width: 100%;
+}
+
+.scheduling-history {
+    margin-bottom: 1em;
+    width: 85%;
+    align-self: center;
+}
+
+.nested-run {
+    margin: 1em;
+}
